airflow/requirements.txt
boto3==1.26.44
finnhub-python==2.4.15
numpy
pandas